refactor: move connection cache to state
